lib: main: Drop use of deprecated sdk/widget.
We get very loud complaints from Firefox (>= 31, at least) if we keep using
the older module.  Furthermore, this makes icons of both 16px and 32px
usable (previously, we were only using the 16px version).

One pending question is: should we keep using an ActionButton or should we
move on to a ToggleButton? The latter seems more appropriate for our use,
but I am keeping it simple right now and making minimal changes.

The documentation can be found [here][0].

[0]: https://developer.mozilla.org/en-US/Add-ons/SDK/High-Level_APIs/ui

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>