Remove content as this is superfluous.
This extension is now superfluous, as new versions of Firefox/Icewease (and
other derived browsers) all support a reader mode that embeds the
readability mode that this extension used to provide.

For historical purposes, though, the old content is still on the `backup`
branch, if anybody wants to look at it.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>